requirements.txt
# Core LangChain Framework
langchain==1.2.6
langchain-core==1.2.7
langchain-community==0.4.1
langchain-chroma==1.1.0
langchain-experimental==0.4.1
langchain-huggingface==1.2.0
langchain-ollama==1.0.1
langgraph==1.0.6

# Document Processing
pypdf==6.6.0

# Environment & Configuration
python-dotenv==1.2.1

# Embeddings & Vector Store
sentence-transformers==5.2.0

# Local LLM Support (Ollama)
ollama==0.6.1

# Additional Dependencies
requests==2.32.5
requests-toolbelt==1.0.0
pyyaml==6.0.2
numpy==2.2.6
scikit-learn==1.8.0

#mem0
mem0ai>=1.0.0

# DeepEval for LLM evaluation
deepeval>=0.21.0

# Core testing framework
pytest>=7.4.0
pytest-html>=4.0.0
pytest-cov>=4.1.0

# Google
google-genai